Adding the message scheduler listener and notification repositories
