+++ microservices/cloud-messaging-service/src/main/kotlin/org/radarbase/appserver/microservices/cloud/messaging/event/listener/quartz/QuartzMessageSchedulerListener.kt
@@ -16,5 +16,279 @@
 
 package org.radarbase.appserver.microservices.cloud.messaging.event.listener.quartz
 
-class QuartzMessageScheduleListener {
+import com.google.common.eventbus.EventBus
+import jakarta.inject.Inject
+import kotlinx.coroutines.runBlocking
+import org.quartz.JobDetail
+import org.quartz.JobKey
+import org.quartz.Scheduler
+import org.quartz.SchedulerException
+import org.quartz.SchedulerListener
+import org.quartz.Trigger
+import org.quartz.TriggerKey
+import org.radarbase.appserver.microservices.core.event.state.MessageState
+import org.radarbase.appserver.microservices.core.event.state.dto.DataMessageStateEventDto
+import org.radarbase.appserver.microservices.core.event.state.dto.NotificationStateEventDto
+import org.radarbase.appserver.microservices.core.repository.DataMessageRepository
+import org.radarbase.appserver.microservices.core.repository.NotificationRepository
+import org.radarbase.appserver.microservices.core.service.quartz.MessageType
+import org.radarbase.appserver.microservices.core.service.quartz.QuartzNamingStrategy
+import org.radarbase.appserver.microservices.core.service.quartz.SimpleQuartzNamingStrategy
+import org.radarbase.jersey.service.AsyncCoroutineService
+import org.slf4j.Logger
+import org.slf4j.LoggerFactory
+import java.time.Instant
+
+class QuartzMessageSchedulerListener @Inject constructor(
+    private val messageStateEventPublisher: EventBus,
+    private val notificationRepository: NotificationRepository,
+    private val dataMessageRepository: DataMessageRepository,
+    private val scheduler: Scheduler,
+    private val asyncService: AsyncCoroutineService,
+) : SchedulerListener {
+    /**
+     * Called by the `[Scheduler]` when a `[JobDetail]` is
+     * scheduled.
+     */
+    override fun jobScheduled(trigger: Trigger) {
+        val jobDetail: JobDetail
+        try {
+            jobDetail = scheduler.getJobDetail(trigger.jobKey)
+        } catch (exc: SchedulerException) {
+            logger.warn(
+                "Encountered error while getting job information from Trigger: ",
+                exc,
+            )
+            return
+        }
+
+        val jobDataMap = jobDetail.jobDataMap
+        val type = MessageType.valueOf(jobDataMap.getString("messageType"))
+        val messageId = jobDataMap.getLongValue("messageId")
+
+        when (type) {
+            MessageType.NOTIFICATION -> {
+                val notification = asyncService.runBlocking {
+                    notificationRepository.find(messageId)
+                } ?: run {
+                    logger.warn("The notification does not exist in database and yet was scheduled.")
+                    return
+                }
+                val notificationStateEvent = NotificationStateEventDto(
+                        notification,
+                        MessageState.SCHEDULED,
+                        null,
+                        Instant.now(),
+                    )
+                messageStateEventPublisher.post(notificationStateEvent)
+            }
+
+            MessageType.DATA -> {
+                val dataMessage = asyncService.runBlocking {
+                    dataMessageRepository.find(messageId)
+                } ?: run {
+                    logger.warn("The data message does not exist in database and yet was scheduled.")
+                    return
+                }
+                val dataMessageStateEvent = DataMessageStateEventDto(
+                    dataMessage,
+                    MessageState.SCHEDULED,
+                    null,
+                    Instant.now(),
+                )
+                messageStateEventPublisher.post(dataMessageStateEvent)
+            }
+
+            MessageType.UNKNOWN -> logger.warn("Job is scheduled for unknown message type")
+        }
+    }
+
+    /**
+     * Called by the `[Scheduler]` when a `[JobDetail]` is
+     * unscheduled.
+     *
+     * @see SchedulerListener.schedulingDataCleared
+     */
+    override fun jobUnscheduled(triggerKey: TriggerKey) {
+        val notificationId: Long
+        try {
+            notificationId = requireNotNull(NAMING_STRATEGY.getMessageId(triggerKey.name)) {
+                "Message ID shouldn't be null when retrieving from naming strategy"
+            }.toLong()
+        } catch (_: NumberFormatException) {
+            logger.warn("The message id could not be established from unscheduled trigger.")
+            return
+        }
+        val notification = runBlocking {
+            notificationRepository.find(notificationId)
+        } ?: run {
+            logger.warn("The notification does not exist in database and yet was unscheduled.")
+            return
+        }
+        val notificationStateEvent = NotificationStateEventDto(
+            notification,
+            MessageState.CANCELLED,
+            null,
+            Instant.now(),
+        )
+        messageStateEventPublisher.post(notificationStateEvent)
+    }
+
+    /**
+     * Called by the `[Scheduler]` when a `[Trigger]` has reached
+     * the condition in which it will never fire again.
+     */
+    override fun triggerFinalized(trigger: Trigger?) {
+        // Not implemented
+    }
+
+    /**
+     * Called by the `[Scheduler]` when a `[Trigger]` has been
+     * paused.
+     */
+    override fun triggerPaused(triggerKey: TriggerKey?) {
+        // Not implemented
+    }
+
+    /**
+     * Called by the `[Scheduler]` when a group of `[Trigger]s` has
+     * been paused.
+     *
+     *
+     * If all groups were paused then triggerGroup will be null
+     *
+     * @param triggerGroup the paused group, or null if all were paused
+     */
+    override fun triggersPaused(triggerGroup: String?) {
+        // Not implemented
+    }
+
+    /**
+     * Called by the `[Scheduler]` when a `[Trigger]` has been
+     * un-paused.
+     */
+    override fun triggerResumed(triggerKey: TriggerKey?) {
+        // Not implemented
+    }
+
+    /**
+     * Called by the `[Scheduler]` when a group of `[Trigger]s` has
+     * been un-paused.
+     */
+    override fun triggersResumed(triggerGroup: String?) {
+        // Not implemented
+    }
+
+    /**
+     * Called by the `[Scheduler]` when a `[JobDetail]` has been
+     * added.
+     */
+    override fun jobAdded(jobDetail: JobDetail?) {
+        // Not implemented
+    }
+
+    /**
+     * Called by the `[Scheduler]` when a `[JobDetail]` has been
+     * deleted.
+     */
+    override fun jobDeleted(jobKey: JobKey?) {
+        // Not implemented
+    }
+
+    /**
+     * Called by the `[Scheduler]` when a `[JobDetail]` has been
+     * paused.
+     */
+    override fun jobPaused(jobKey: JobKey?) {
+        // Not implemented
+    }
+
+    /**
+     * Called by the `[Scheduler]` when a group of `[JobDetail]s`
+     * has been paused.
+     *
+     * @param jobGroup the paused group, or null if all were paused
+     */
+    override fun jobsPaused(jobGroup: String?) {
+        // Not implemented
+    }
+
+    /**
+     * Called by the `[Scheduler]` when a `[JobDetail]` has been
+     * un-paused.
+     */
+    override fun jobResumed(jobKey: JobKey?) {
+        // Not implemented
+    }
+
+    /**
+     * Called by the `[Scheduler]` when a group of `[JobDetail]s`
+     * has been un-paused.
+     */
+    override fun jobsResumed(jobGroup: String?) {
+        // Not implemented
+    }
+
+    /**
+     * Called by the `[Scheduler]` when a serious error has occurred within the
+     * scheduler - such as repeated failures in the `JobStore`, or the inability to
+     * instantiate a `Job` instance when its `Trigger` has fired.
+     *
+     *
+     * The `getErrorCode()` method of the given SchedulerException can be used to
+     * determine more specific information about the type of error that was encountered.
+     */
+    override fun schedulerError(msg: String?, cause: SchedulerException?) {
+        // Not implemented
+    }
+
+    /**
+     * Called by the `[Scheduler]` to inform the listener that it has move to standby
+     * mode.
+     */
+    override fun schedulerInStandbyMode() {
+        // Not implemented
+    }
+
+    /**
+     * Called by the `[Scheduler]` to inform the listener that it has started.
+     */
+    override fun schedulerStarted() {
+        // Not implemented
+    }
+
+    /**
+     * Called by the `[Scheduler]` to inform the listener that it is starting.
+     */
+    override fun schedulerStarting() {
+        // Not implemented
+    }
+
+    /**
+     * Called by the `[Scheduler]` to inform the listener that it has shutdown.
+     */
+    override fun schedulerShutdown() {
+        // Not implemented
+    }
+
+    /**
+     * Called by the `[Scheduler]` to inform the listener that it has begun the
+     * shutdown sequence.
+     */
+    override fun schedulerShuttingdown() {
+        // Not implemented
+    }
+
+    /**
+     * Called by the `[Scheduler]` to inform the listener that all jobs, triggers and
+     * calendars were deleted.
+     */
+    override fun schedulingDataCleared() {
+        // Not implemented
+    }
+
+    companion object {
+        private val NAMING_STRATEGY: QuartzNamingStrategy = SimpleQuartzNamingStrategy()
+        private val logger: Logger = LoggerFactory.getLogger(QuartzMessageSchedulerListener::class.java)
+    }
 }
